+++ packages/wdio-config/src/node/ConfigParser.ts
@@ -6,10 +6,8 @@ import type { Capabilities, Options, Services } from '@wdio/types'
 
 import RequireLibrary from './RequireLibrary.js'
 import FileSystemPathService from './FileSystemPathService.js'
-import {
-    removeLineNumbers, isCucumberFeatureWithLineNumber, validObjectOrArray,
-    loadAutoCompilers, makeRelativeToCWD
-} from '../utils.js'
+import { makeRelativeToCWD, loadAutoCompilers } from './utils.js'
+import { removeLineNumbers, isCucumberFeatureWithLineNumber, validObjectOrArray } from '../utils.js'
 import { SUPPORTED_HOOKS, SUPPORTED_FILE_EXTENSIONS, DEFAULT_CONFIGS, NO_NAMED_CONFIG_EXPORT } from '../constants.js'
 
 import type { PathService, ModuleImportService } from '../types.js'
@@ -174,7 +172,8 @@ export default class ConfigParser {
          */
         if (object.specs && object.specs.length > 0) {
             this._config.specs = object.specs as string[]
-        } else if (object.exclude && object.exclude.length > 0) {
+        }
+        if (object.exclude && object.exclude.length > 0) {
             this._config.exclude = object.exclude as string[]
         }