+++ packages/wdio-config/tests/__fixtures__/wdio.conf.rdc.ts
@@ -8,7 +8,7 @@ export const config: Options.Testrunner = {
     key: '50fa142c-3121-4gb0-9p07-8q326vvbq7b0',
     specs: [path.join(TEST_ROOT, '*.test.ts')],
     exclude: [
-        path.join(TEST_ROOT, '/RequireLibrary.test.ts'),
+        path.join(TEST_ROOT, '/FileSystemPathService.test.ts'),
         path.join(TEST_ROOT, '/validateConfig.test.ts')
     ],
     capabilities: [{
@@ -16,8 +16,8 @@ export const config: Options.Testrunner = {
         testobject_api_key: '1',
     }],
     suites: {
-        unit: [path.join(TEST_ROOT, 'configparser.test.ts')],
-        mobile: [path.join(TEST_ROOT, 'RequireLibrary.test.ts')],
+        unit: [path.join(TEST_ROOT, 'node', 'configparser.test.ts')],
+        mobile: [path.join(TEST_ROOT, 'node', 'FileSystemPathService.test.ts')],
         functional: [
             path.join(TEST_ROOT, 'validateConfig.test.ts'),
             path.join(TEST_ROOT, '..', 'src/index.ts')