+++ packages/wdio-utils/src/utils.ts
@@ -1,13 +1,22 @@
-import fs from 'node:fs'
-import path from 'node:path'
-import { pathToFileURL } from 'node:url'
-import { resolve } from 'import-meta-resolve'
+import type { Options, Services, Clients } from '@wdio/types'
 
-import type { Services, Clients } from '@wdio/types'
+import { SUPPORTED_BROWSERNAMES, DEFAULT_PROTOCOL, DEFAULT_HOSTNAME, DEFAULT_PATH } from './constants.js'
 
 const SCREENSHOT_REPLACEMENT = '"<Screenshot[base64]>"'
 const SCRIPT_PLACEHOLDER = '"<Script[base64]>"'
 const REGEX_SCRIPT_NAME = /return \((async )?function (\w+)/
+const SLASH = '/'
+
+function assertPath(path?: any) {
+    if (typeof path !== 'string') {
+        throw new TypeError('Path must be a string. Received ' + JSON.stringify(path))
+    }
+}
+
+export function isAbsolute(p: string) {
+    assertPath(p)
+    return p.length > 0 && p.charCodeAt(0) === SLASH.codePointAt(0)
+}
 
 /**
  * overwrite native element commands with user defined
@@ -181,17 +190,22 @@ export async function safeImport (name: string): Promise<Services.ServicePlugin 
          * then we also need to search in the project, we do that by defining
          * 'localNodeModules' and searching from that also.
          *
-         * Note that import-meta-resolve will resolve to CJS no ESM export is found
+         * Note that import-meta-resolve will resolve to CJS no ESM export is found.
+         * Only in Node.js environments
          */
-        const localNodeModules = path.join(process.cwd(), 'node_modules')
-
-        try {
-            importPath = await resolve(name, import.meta.url)
-        } catch (err: any) {
+        if (!globalThis.window) {
+            const { resolve } = await import('import-meta-resolve')
             try {
-                importPath = await resolve(name, pathToFileURL(localNodeModules).toString())
+                importPath = await resolve(name, import.meta.url)
             } catch (err: any) {
-                return null
+                const { join } = await import('node:path')
+                const { pathToFileURL } = await import('node:url')
+                const localNodeModules = join(process.cwd(), 'node_modules')
+                try {
+                    importPath = await resolve(name, pathToFileURL(localNodeModules).toString())
+                } catch (err: any) {
+                    return null
+                }
             }
         }
     } catch (err: any) {
@@ -265,24 +279,6 @@ export function isBase64(str: string) {
 }
 
 /**
- * Helper utility to check file access
- * @param {string} file file to check access for
- * @return              true if file can be accessed
- */
-export const canAccess = (file?: string) => {
-    if (!file) {
-        return false
-    }
-
-    try {
-        fs.accessSync(file)
-        return true
-    } catch (err: any) {
-        return false
-    }
-}
-
-/**
  * sleep
  * @param {number=0} ms number in ms to sleep
  */
@@ -300,4 +296,35 @@ export function isAppiumCapability(caps: WebdriverIO.Capabilities): boolean {
         // @ts-expect-error outdated jsonwp cap
         (caps.automationName || caps['appium:automationName'] || caps.deviceName || caps.appiumVersion)
     )
+}
+
+/**
+ * helper method to determine if we need to setup a browser driver
+ * which is:
+ *   - whenever the user has set connection options that differ
+ *     from the default, or a port is set
+ *   - whenever the user defines `user` and `key` which later will
+ *     update the connection options
+ */
+export function definesRemoteDriver(options: Pick<Options.WebDriver, 'user' | 'key' | 'protocol' | 'hostname' | 'port' | 'path'>) {
+    return Boolean(
+        (options.protocol && options.protocol !== DEFAULT_PROTOCOL) ||
+        (options.hostname && options.hostname !== DEFAULT_HOSTNAME) ||
+        Boolean(options.port) ||
+        (options.path && options.path !== DEFAULT_PATH) ||
+        Boolean(options.user && options.key)
+    )
+}
+
+export function isChrome (browserName?: string) {
+    return Boolean(browserName && SUPPORTED_BROWSERNAMES.chrome.includes(browserName.toLowerCase()))
+}
+export function isSafari (browserName?: string) {
+    return Boolean(browserName && SUPPORTED_BROWSERNAMES.safari.includes(browserName.toLowerCase()))
+}
+export function isFirefox (browserName?: string) {
+    return Boolean(browserName && SUPPORTED_BROWSERNAMES.firefox.includes(browserName.toLowerCase()))
+}
+export function isEdge (browserName?: string) {
+    return Boolean(browserName && SUPPORTED_BROWSERNAMES.edge.includes(browserName.toLowerCase()))
 }