+++ packages/wdio-utils/tests/node/utils.test.ts
@@ -2,14 +2,19 @@ import os from 'node:os'
 import path from 'node:path'
 import url from 'node:url'
 import cp from 'node:child_process'
-import type fs from 'node:fs'
+import fs from 'node:fs'
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { canDownload, resolveBuildId, detectBrowserPlatform } from '@puppeteer/browsers'
-import { locateChrome } from 'locate-app'
+import { locateChrome, locateApp } from 'locate-app'
 
-import { parseParams, getBuildIdByChromePath, getBuildIdByFirefoxPath, setupPuppeteerBrowser, definesRemoteDriver } from '../../src/driver/utils.js'
+import {
+    parseParams, getBuildIdByChromePath, getBuildIdByFirefoxPath, setupPuppeteerBrowser,
+    canAccess
+} from '../../src/node/utils.js'
 
 const __dirname = path.dirname(url.fileURLToPath(import.meta.url))
+
+vi.mock('fs', () => import(path.join(process.cwd(), '__mocks__', 'fs')))
 
 vi.mock('node:os', () => ({
     default: {
@@ -20,11 +25,13 @@ vi.mock('node:os', () => ({
 
 vi.mock('locate-app', () => ({
     locateChrome: vi.fn().mockResolvedValue('/path/to/chrome'),
-    locateFirefox: vi.fn().mockResolvedValue('/path/to/firefox')
+    locateFirefox: vi.fn().mockResolvedValue('/path/to/firefox'),
+    locateApp: vi.fn().mockResolvedValue('/path/to/chromium')
 }))
 
 vi.mock('node:fs', () => ({
     default: {
+        accessSync: vi.fn(),
         readdirSync: vi.fn().mockReturnValue([
             '114.0.5735.199',
             '115.0.5790.110',
@@ -57,7 +64,7 @@ vi.mock('node:child_process', () => ({
 }))
 
 vi.mock('@puppeteer/browsers', () => ({
-    Browser: { CHROME: 'chrome', FIREFOX: 'firefox' },
+    Browser: { CHROME: 'chrome', FIREFOX: 'firefox', CHROMIUM: 'chromium' },
     ChromeReleaseChannel: { STABLE: 'stable' },
     detectBrowserPlatform: vi.fn(),
     resolveBuildId: vi.fn().mockReturnValue('116.0.5845.110'),
@@ -115,7 +122,6 @@ describe('driver utils', () => {
         })
 
         it('should do nothing if browser binary is defined within caps', async () => {
-            vi.mocked(detectBrowserPlatform).mockReturnValueOnce('mac' as any)
             await expect(setupPuppeteerBrowser('/foo/bar', {
                 'goog:chromeOptions': { binary: '/my/chrome' }
             })).resolves.toEqual({
@@ -138,6 +144,22 @@ describe('driver utils', () => {
                 browserVersion: '116.0.5845.110',
                 executablePath: '/path/to/stable'
             })
+            expect(resolveBuildId).toBeCalledTimes(1)
+            expect(resolveBuildId).toBeCalledWith('chrome', 'windows', '116.0.5845.110')
+        })
+
+        it('should look for Chromium browser if defined as browser name', async () => {
+            vi.mocked(detectBrowserPlatform).mockReturnValueOnce('windows' as any)
+            vi.mocked(locateApp).mockResolvedValue('')
+            const caps = { browserName: 'chromium' }
+            await expect(setupPuppeteerBrowser('/foo/bar', caps)).resolves.toEqual( {
+                browserVersion: '116.0.5845.110',
+                executablePath: '/foo/bar/executable'
+            })
+            expect(caps.browserName).toBe('chrome')
+            expect(resolveBuildId).toBeCalledTimes(2)
+            expect(resolveBuildId).toBeCalledWith('chromium', 'windows', 'latest')
+            expect(resolveBuildId).toBeCalledWith('chrome', 'windows', 'latest')
         })
 
         it('should throw if browser version is not found', async () => {
@@ -157,15 +179,17 @@ describe('driver utils', () => {
             expect(resolveBuildId).toBeCalledWith('chrome', 'windows', '1.2.3')
         })
     })
+})
 
-    it('definesRemoteDriver', () => {
-        expect(definesRemoteDriver({})).toBe(false)
-        expect(definesRemoteDriver({ hostname: 'foo' })).toBe(true)
-        expect(definesRemoteDriver({ port: 1 })).toBe(true)
-        expect(definesRemoteDriver({ path: 'foo' })).toBe(true)
-        expect(definesRemoteDriver({ protocol: 'foo' })).toBe(true)
-        expect(definesRemoteDriver({ user: 'foo' })).toBe(false)
-        expect(definesRemoteDriver({ key: 'foo' })).toBe(false)
-        expect(definesRemoteDriver({ user: 'foo', key: 'bar' })).toBe(true)
+describe('utils:canAccess', () => {
+    it('canAccess', () => {
+        expect(canAccess('/foobar')).toBe(true)
+        expect(fs.accessSync).toBeCalledWith('/foobar')
+
+        // @ts-ignore
+        fs.accessSync.mockImplementation(() => {
+            throw new Error('upps')
+        })
+        expect(canAccess('/foobar')).toBe(false)
     })
 })